RSG: Edit and expand section on indentation (WIP)
In the Rustic Style Guide post, edit the section on indentation,
including expanding it with more code examples and partially
standardsese-ifying it.

There is still more to be done for this section, and plenty more to be
done for the other sections. Most or all of the things to be done in
this section have been noted with "TODO" comments.

I have been marking up code examples as figures; these will most
likely be aesthetically suboptimal until I write appropriate CSS for
them.