RSG: Edit and expand function headers section (WIP)
In the Rustic Style Guide post, edit and expand the section on
function headers, albeit not as thoroughly as the section on
indentation.

This editing includes renaming the section from "Function Definitions"
to "Function Headers", because it doesn't appear to be about any parts
of a function definition other than the header.

I've left three TODO items here.