RSG: Edit and expand start of § Function Body
In the Rustic Style Guide post, edit and expand the beginning of the
section on the formatting of the contents of the bodies of functions.

Much of what I've written here, I've extrapolated from the code
examples rather than from the normative English text; I advise that
ubsan make sure I didn't introduce any guidance contrary to her
intentions.